+++ src/components/Home.js
@@ -7,7 +7,7 @@ import { BrowserRouter as Router, Route } from 'react-router-dom';
 import Logger from '../js/Logger';
 import Menubar from './MenuBar';
 import BlankComponent from '@gitobi/react-blank-component';
-import DevicesWaterings from './containers/DevicesWaterings';
+import DevicesWaterings from '../containers/DevicesWaterings';
 
 class Home extends Component {
   constructor(props) {
